4/pyproject.toml
[tool.poetry]
name = "exercise"
version = "0.1.0"
description = "exercise"
authors = ["exercise"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
